telegram_bot/requirements.txt
python-telegram-bot==13.7
openai==0.27.0
SQLAlchemy==1.4.22
Flask==2.0.1
python-dotenv==0.19.1
psycopg2-binary==2.9.1